Act on Kodi debug logging
If Kodi was configured to get debug logging, ignore the addon log level
altogether and log as DEBUG.

Since we log when the service is stopped, I opted to also log when the
service is started. Otherwise it is weird if you disable and enable it,
it only logged that the service was stopped, but not that it was started
again.