customAutoPlay and Skip-to-End improvements
This PR includes:
- Enable customAutoPlayTime by default which avoids skip-to-end for
  most cases
- Only skip-to-end when we are outside of the default Kodi margin (95%)
- Make sure we play the next episode, also when getTime() or
  getTotalTime() fail
- In the unlikely case we do need seekTime(), do not seek to the end,
  but stay within margin

First of all, skipping the credits is not needed in the majority of cases,
because the pop-up appears within the 95% default margin in Kodi.

Secondly, this seeking, when done using VOD add-ons, causes the player
to exit before the next items is being played, so it doesn't give a nice
user experience.

By skipping (the end credits) to 59/60th before the end, rather
than the actual end, we avoid issues with the Kodi Player.

- Seeking to the end is not exact (may go past the end)
- Avoid any Player events that interfere with us playing the next item
  in the playlist

We subtract 1 second to compensate for the inaccuracy of getTotalTime().

And we use 59/60th so that for a 1 minute video it would seek to 1 sec
before the end. And for a 1 hour video it would seek to 1 min before the
end. This would be far withing the default 95% margin of Kodi.