Implementation of shuffle mode
When clicking the shuffle button on the UpNext popup, a random episode from the current tvshow will be played, instead of the next episode. Random episodes from the current tvshow will continue to be selected by UpNex, rather than the next episode, until the shuffle button is clicked again and shuffle mode is turned off.