Implement persistent storage of video frame hashes
Hashes are stored as a file on playback of a video file where end credits are not detected. File is named upnext.json and is saved in the same directory where the episode is located.

Playback of any episode from the same directory where the upnext.json file is located will compare the current video frame hash with the stored hashes to determine whether the video frames match, which is then used to detect the end credits.

TODO:
[1] Fix known problem where playback of video file used to generate and store the image hashes will always trigger end credits detection. Need to compare with stored hashes from a different video file in the same directory instead.
[2] Store individual episode end credit times after they are detected, rather than always trying to detect the end credits on every playback. Alternatively look into creating EDL files.