Remove comparison of current and next episodeid in watched episode check
This was a remnant of code from the very first release of UpNext.

It is no longer possible for the associated code path to execute when the current and next episodeid is the same, unless this has specifically been requested i.e. playing a playlist with the same file in consecutive positions.

This check is being removed as it serves no other purpose now, and prevents proper playlist handling.