Addresses https://github.com/xbmc/xbmc/issues/19378
Video mediatype may be discarded when a listItem VideoInfoTag is merged.

This can prevent UpNext from working with an addon, even if the addon sends the data required for UpNext to function, as UpNext checks the videoplayer.content(episodes) infobool for all video content, when this is only actually needed for library content.

Note that this should also be fixed for playlist content, however that change is a little bit more complicated and requires more significant modifications.