Remove check for watched episode in playbackmanager
Logic was incorrect - was looking at currently playing episode playcount rather than next episode so would never work correctly.

Also this check would only work for playlist or plugin content as the request for next episode in the library already filters out watched episodes.

With 6b56d41 the request for next playlist item also filters out watched episodes, so this check is now removed as it is only applicable to addon content.

For plugin content it is not possible to skip to next unwatched automatically without integration with the plugin player callback