Fix for significant slowdown of detector over time
Was previously comparing each hash to as many as 240 hashes, for every episode in a season that had been played, which quickly leads to comparing each hash to thousands of other hashes.
Now limited to only 20 hashes for a maximum of two episodes in a season.
Also remove backwards compatibility with old hash file format